+++ packages/dialect-react-sdk/hooks/internal/swrCache.ts
@@ -44,3 +44,11 @@ export const WALLET_DAPP_ADDRESSES_CACHE_KEY_FN = (
   wallet?.publicKey?.toBase58() +
   '_' +
   dappPublicKey?.toBase58();
+
+export const WALLET_NOTIFICATION_SUBSCRIPTIONS_CACHE_KEY_FN = (
+  wallet: Wallets
+) => 'WALLET_NOTIFICATION_SUBSCRIPTIONS_' + wallet?.publicKey?.toBase58();
+
+export const DAPP_NOTIFICATION_SUBSCRIPTIONS_CACHE_KEY_FN = (
+  dapp: Dapp | null
+) => 'DAPP_NOTIFICATION_SUBSCRIPTIONS_' + dapp?.publicKey?.toBase58();