+++ packages/dialect-react-ui/components/Broadcast/BroadcastForm.tsx
@@ -2,11 +2,13 @@ import {
   AddressType,
   Dapp,
   DappAddress,
+  DappNotificationSubscription,
   useDappAddresses,
+  useDappNotificationSubscriptions,
 } from '@dialectlabs/react-sdk';
 import clsx from 'clsx';
-import { useMemo, useState } from 'react';
-import { Button, ValueRow } from '../common';
+import { ReactNode, useEffect, useMemo, useState } from 'react';
+import { Button, Loader, ValueRow } from '../common';
 import { H1, Input, P, Textarea } from '../common/preflighted';
 import { useTheme } from '../common/providers/DialectThemeProvider';
 import ToastMessage from '../common/ToastMessage';
@@ -14,36 +16,68 @@ import ToastMessage from '../common/ToastMessage';
 // utf8 bytes
 const MESSAGE_BYTES_LIMIT = 800;
 const TITLE_BYTES_LIMIT = 100;
+const ADDRESSES_REFRESH_INTERVAL = 10000;
 
 interface BroadcastFormProps {
   dapp: Dapp;
 }
 
-const getUserCount = (addresses: DappAddress[]) => {
-  // Users count = set of unique wallets, associated with enabled dapp addresses, associated with verified addresses
-  const enabledAndVerified = addresses
+const getUsersCount = (
+  addresses: DappAddress[],
+  subscriptions: DappNotificationSubscription[],
+  notificationTypeId?: string | null,
+  addressTypePredicate: (addressType: AddressType) => boolean = () => true
+) => {
+  const enabledAndVerifiedAddresses = addresses
     .filter((address) => address.enabled)
     .filter((address) => address.address.verified)
-    .map((address) => address.address.wallet.publicKey.toBase58());
-  return [...new Set(enabledAndVerified)].length;
+    .filter((address) => addressTypePredicate(address.address.type));
+  const enabledSubscriptionsPKs = subscriptions
+    .filter(
+      (sub) =>
+        (sub.notificationType.id === notificationTypeId ||
+          !notificationTypeId) &&
+        sub.subscriptions.find((subscription) => subscription.config.enabled)
+    )
+    .flatMap((it) => it.subscriptions);
+  const filtered = enabledSubscriptionsPKs.filter((subscription) =>
+    enabledAndVerifiedAddresses.find((address) =>
+      address.address.wallet.publicKey.equals(subscription.wallet.publicKey)
+    )
+  );
+  return [...new Set(filtered.map((it) => it.wallet.publicKey.toBase58()))]
+    .length;
 };
 
-const getAddressesCounts = (addresses: DappAddress[]) => {
-  const enabledAndVerified = addresses
-    .filter((address) => address.enabled)
-    .filter((address) => address.address.verified);
-  const wallets = enabledAndVerified.filter(
-    (address) => address.address.type === AddressType.Wallet
-  ).length;
-  const emails = enabledAndVerified.filter(
-    (address) => address.address.type === AddressType.Email
-  ).length;
-  const phones = enabledAndVerified.filter(
-    (addresses) => addresses.address.type === AddressType.PhoneNumber
-  ).length;
-  const telegrams = enabledAndVerified.filter(
-    (address) => address.address.type === AddressType.Telegram
-  ).length;
+const getAddressesCounts = (
+  addresses: DappAddress[],
+  subscriptions: DappNotificationSubscription[],
+  notificationTypeId?: string | null
+) => {
+  const wallets = getUsersCount(
+    addresses,
+    subscriptions,
+    notificationTypeId,
+    (it) => it === AddressType.Wallet
+  );
+  const emails = getUsersCount(
+    addresses,
+    subscriptions,
+    notificationTypeId,
+    (it) => it === AddressType.Email
+  );
+  const phones = getUsersCount(
+    addresses,
+    subscriptions,
+    notificationTypeId,
+    (it) => it === AddressType.PhoneNumber
+  );
+  const telegrams = getUsersCount(
+    addresses,
+    subscriptions,
+    notificationTypeId,
+    (it) => it === AddressType.Telegram
+  );
   return {
     wallets,
     emails,
@@ -52,8 +86,16 @@ const getAddressesCounts = (addresses: DappAddress[]) => {
   };
 };
 
-const getAddressesSummary = (addresses: DappAddress[]) => {
-  const { wallets, emails, phones, telegrams } = getAddressesCounts(addresses);
+const getAddressesSummary = (
+  addresses: DappAddress[],
+  subscriptions: DappNotificationSubscription[],
+  notificationTypeId?: string | null
+) => {
+  const { wallets, emails, phones, telegrams } = getAddressesCounts(
+    addresses,
+    subscriptions,
+    notificationTypeId
+  );
   return [
     wallets && `${wallets} wallet${wallets > 1 ? 's' : ''} (off-chain)`,
     emails && `${emails} email${emails > 1 ? 's' : ''}`,
@@ -65,7 +107,17 @@ const getAddressesSummary = (addresses: DappAddress[]) => {
 };
 
 function BroadcastForm({ dapp }: BroadcastFormProps) {
-  const { addresses, isFetching: isFetchingAddresses } = useDappAddresses();
+  const {
+    subscriptions: notificationsSubscriptions,
+    // isFetching: isFetchingNotificationSubscriptions,
+    // errorFetching: errorFetchingNotificationSubscriptions,
+  } = useDappNotificationSubscriptions();
+  const [notificationTypeId, setNotificationTypeId] = useState<string | null>(
+    notificationsSubscriptions[0]?.notificationType.id ?? null
+  );
+  const { addresses, isFetching: isFetchingAddresses } = useDappAddresses({
+    refreshInterval: ADDRESSES_REFRESH_INTERVAL,
+  });
   const { textStyles, colors, outlinedInput } = useTheme();
   // Consider moving error handling to the useDapp context
   const [error, setError] = useState<Error | null>(null);
@@ -83,10 +135,28 @@ function BroadcastForm({ dapp }: BroadcastFormProps) {
     [textEncoder, message]
   );
 
-  const usersCount = useMemo(() => getUserCount(addresses), [addresses]);
+  useEffect(() => {
+    !notificationTypeId &&
+      notificationsSubscriptions.length > 0 &&
+      setNotificationTypeId(
+        notificationsSubscriptions[0]?.notificationType.id ?? null
+      );
+  }, [notificationTypeId, notificationsSubscriptions]);
+
+  const usersCount = useMemo(
+    () =>
+      getUsersCount(addresses, notificationsSubscriptions, notificationTypeId),
+    [addresses, notificationTypeId, notificationsSubscriptions]
+  );
+
   const addressesSummary = useMemo(
-    () => getAddressesSummary(addresses),
-    [addresses]
+    () =>
+      getAddressesSummary(
+        addresses,
+        notificationsSubscriptions,
+        notificationTypeId
+      ),
+    [addresses, notificationsSubscriptions, notificationTypeId]
   );
   const noUsers = usersCount === 0;
   const isSubmitDisabled =
@@ -95,18 +165,22 @@ function BroadcastForm({ dapp }: BroadcastFormProps) {
     messageLength > MESSAGE_BYTES_LIMIT ||
     titleLength > TITLE_BYTES_LIMIT ||
     noUsers;
-  let usersString = `${usersCount} user${usersCount > 1 ? 's' : ''}`;
+  let usersInfo: ReactNode = `${usersCount} user${usersCount > 1 ? 's' : ''}`;
 
   if (isFetchingAddresses) {
-    usersString = 'Loading users...';
+    usersInfo = <Loader />;
   } else if (noUsers) {
-    usersString = 'No users yet';
+    usersInfo = 'No users yet';
   }
 
   const sendBroadcastMessage = async () => {
     setIsSending(true);
     try {
-      await dapp.messages.send({ title, message });
+      await dapp.messages.send({
+        title,
+        message,
+        ...(notificationTypeId && { notificationTypeId }),
+      });
       setTitle('');
       setMessage('');
       setError(null);
@@ -120,13 +194,41 @@ function BroadcastForm({ dapp }: BroadcastFormProps) {
     }
   };
 
+  const renderNotificationTypeSelect = () => {
+    if (!notificationsSubscriptions.length || !notificationTypeId) {
+      return '📢 Broadcast';
+    }
+
+    if (notificationsSubscriptions.length === 1) {
+      return notificationsSubscriptions[0]?.notificationType.name;
+    }
+
+    return (
+      // TODO: create a preflighted version of select with :focus-visible and other default things, which is already configured for inputs and buttons
+      <select
+        value={notificationTypeId}
+        className="dt-bg-transparent dt-text-inherit focus:dt-outline-0 dt-text-right"
+        onChange={(event) => setNotificationTypeId(event.target.value)}
+      >
+        {notificationsSubscriptions.map(({ notificationType }) => (
+          <option key={notificationType.id} value={notificationType.id}>
+            {notificationType.name}
+          </option>
+        ))}
+      </select>
+    );
+  };
+
   return (
     <div className="dt-flex dt-flex-col dt-space-y-2">
       <H1 className={clsx(textStyles.h1, colors.primary, 'dt-mb-4')}>
         Create broadcast
       </H1>
-      <ValueRow label="📢 Broadcast" className="dt-w-full">
-        <span title={addressesSummary}>{usersString}</span>
+      <ValueRow label="Category" className="dt-w-full">
+        <span>{renderNotificationTypeSelect()}</span>
+      </ValueRow>
+      <ValueRow label="📢 Broadcast users coverage" className="dt-w-full">
+        <span title={addressesSummary}>{usersInfo}</span>
       </ValueRow>
       <div>
         <Input