+++ packages/dialect-react-ui/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [UNRELEASED]
 
+- feature: configurable notifications in `<NotificationsButton/>` component settings
+- feature: configurable notifications in `<BroadcastForm/>` component
+
 ## [1.0.0-beta.21] - 2022-07-25
 - chore: minor stylistic changes
 - fix: add missing notifs header